# Basic auth alloc