## Description

Fix for test basic header as described in RFC 7617, like

Authorization: Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==

Also avoid string buffer alloc in `strings.ToLower` if  string not in lower case, like `Basic`

Fixes # (issue)

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)