@msaf1980 2023-02-13
Before (current master)

```
$ go test -benchmem -run=^$ -bench ^Benchmark_Middleware_BasicAuth github.com/gofiber/fiber/v2/middleware/basicauth
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/basicauth
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
Benchmark_Middleware_BasicAuth-6         	 3374157	       356.2 ns/op	      80 B/op	       5 allocs/op
Benchmark_Middleware_BasicAuth_Upper-6   	 3117049	       387.5 ns/op	      80 B/op	       6 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2/middleware/basicauth	3.168s
```

After (this PR)

```
$ go test -benchmem -run=^$ -bench ^Benchmark_Middleware_BasicAuth github.com/gofiber/fiber/v2/middleware/basicauth
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/basicauth
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
Benchmark_Middleware_BasicAuth-6         	 3175054	       371.6 ns/op	      80 B/op	       5 allocs/op
Benchmark_Middleware_BasicAuth_Upper-6   	 3207657	       376.4 ns/op	      80 B/op	       5 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2/middleware/basicauth	3.156s
```

As I see, if auth header like `Authorization: basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==` we have some perf drawback.
But if request like `Authorization: Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==` (as described in RFC, and most requests is) - we have minus one allocation.