@efectn 2023-02-13
Can you use https://github.com/gofiber/fiber/blob/master/utils/strings.go#L65 and send results again