@msaf1980 2023-02-13
With utils.EqualFold

$ go test -benchmem -run=^$ -bench ^Benchmark_Middleware_BasicAuth github.com/gofiber/fiber/v2/middleware/basicauth
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/basicauth
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
Benchmark_Middleware_BasicAuth-6         	 3487176	       343.3 ns/op	      80 B/op	       5 allocs/op
Benchmark_Middleware_BasicAuth_Upper-6   	 3503778	       348.3 ns/op	      80 B/op	       5 allocs/op

With utils.ToLower

$ go test -benchmem -run=^$ -bench ^Benchmark_Middleware_BasicAuth github.com/gofiber/fiber/v2/middleware/basicauth
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/basicauth
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
Benchmark_Middleware_BasicAuth-6         	 3373893	       358.1 ns/op	      80 B/op	       6 allocs/op
Benchmark_Middleware_BasicAuth_Upper-6   	 3374263	       356.8 ns/op	      80 B/op	       6 allocs/op

Update PR for use utils.EqualFold ?