@efectn 2023-02-13
> > sure
> 
> Done. If you plan to merge, may be better squash commit on merge ?

Don't need it. We're already merging squashed commits into the master